# Optimized packet load in gemm_pack_rhs (Packet16f/8f) for CuboidConv