When packet is split across 2 adjacent rows (on same column), we read it as two
'partial' packets and then merge these into 1. This optimization is similar to
that implemented in SpatialConvolution (commid id:
e048543b0543ab12fa6e74537a21227a0837fcd1)

This optimization shows significant speedup in CuboidalConvolution with certain
parameters. Numbers from tensorflow/core/kernels:eigen_benchmark_cpu_test are
below.

For AVX512, the patch speeds up 2 (out of 4) instances and leaves the other 2 as
is. For AVX2, the patch speeds up 1 (out of 4) instances and leaves the other 3
as is.

AVX512

Parameters                                                 | Runtime without patch (ns) | Runtime with patch (ns) | Speedup
-----------------------------------------------------------|----------------------------|-------------------------|---------
BM_CUBOID_NAME(CuboidConvolution,2,8,25,25,25,4,16,5,5,5)  |         321454780          |         133686280       |  2.40X
BM_CUBOID_NAME(CuboidConvolution,4,8,25,25,25,4,16,5,5,5)  |         160635090          |          67167930       |  2.39X
BM_CUBOID_NAME(CuboidConvolution,8,8,25,25,25,4,16,5,5,5)  |          79329930          |          33715910       |  2.35X
BM_CUBOID_NAME(CuboidConvolution,16,8,25,25,25,4,16,5,5,5) |          50762340          |          21891730       |  2.32X
BM_CUBOID_NAME(CuboidConvolution,2,8,25,25,25,8,16,5,5,5)  |         287244580          |         162654990       |  1.77X
BM_CUBOID_NAME(CuboidConvolution,4,8,25,25,25,8,16,5,5,5)  |         144165180          |          81628980       |  1.77X
BM_CUBOID_NAME(CuboidConvolution,8,8,25,25,25,8,16,5,5,5)  |          72296380          |          41090350       |  1.76X
BM_CUBOID_NAME(CuboidConvolution,16,8,25,25,25,8,16,5,5,5) |          46290810          |          27015430       |  1.71X
BM_CUBOID_NAME(CuboidConvolution,2,2,9,31,31,64,64,5,5,5)  |          87803480          |          87431440       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,4,2,9,31,31,64,64,5,5,5)  |          44564890          |          44420390       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,8,2,9,31,31,64,64,5,5,5)  |          22709950          |          22677270       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,16,2,9,31,31,64,64,5,5,5) |          15299040          |          15405270       |  0.99X
BM_CUBOID_NAME(CuboidConvolution,2,2,5,27,27,64,64,5,5,5)  |          37558060          |          37290600       |  1.01X
BM_CUBOID_NAME(CuboidConvolution,4,2,5,27,27,64,64,5,5,5)  |          19069400          |          19132680       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,8,2,5,27,27,64,64,5,5,5)  |          10088400          |          10110880       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,16,2,5,27,27,64,64,5,5,5) |           6913840          |           6807780       |  1.02X

AVX2

Parameters                                                 | Runtime without patch (ns) | Runtime with patch (ns) | Speedup
-----------------------------------------------------------|----------------------------|-------------------------|---------
BM_CUBOID_NAME(CuboidConvolution,2,8,25,25,25,4,16,5,5,5)  |         166535650          |         115406110       |  1.44X
BM_CUBOID_NAME(CuboidConvolution,4,8,25,25,25,4,16,5,5,5)  |          83449610          |          57959210       |  1.44X
BM_CUBOID_NAME(CuboidConvolution,8,8,25,25,25,4,16,5,5,5)  |          41960720          |          29059400       |  1.44X
BM_CUBOID_NAME(CuboidConvolution,16,8,25,25,25,4,16,5,5,5) |          27584140          |          19657640       |  1.40X
BM_CUBOID_NAME(CuboidConvolution,2,8,25,25,25,8,16,5,5,5)  |          62304280          |          62090900       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,4,8,25,25,25,8,16,5,5,5)  |          31291390          |          31224440       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,8,8,25,25,25,8,16,5,5,5)  |          15775900          |          15744130       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,16,8,25,25,25,8,16,5,5,5) |          10358790          |          10637100       |  0.97X
BM_CUBOID_NAME(CuboidConvolution,2,2,9,31,31,64,64,5,5,5)  |         126984890          |         126675250       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,4,2,9,31,31,64,64,5,5,5)  |          63888330          |          63775570       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,8,2,9,31,31,64,64,5,5,5)  |          32098130          |          31990920       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,16,2,9,31,31,64,64,5,5,5) |          21573360          |          21775700       |  0.99X
BM_CUBOID_NAME(CuboidConvolution,2,2,5,27,27,64,64,5,5,5)  |          53339690          |          53429810       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,4,2,5,27,27,64,64,5,5,5)  |          26970450          |          26968820       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,8,2,5,27,27,64,64,5,5,5)  |          13980030          |          13981490       |  1.00X
BM_CUBOID_NAME(CuboidConvolution,16,2,5,27,27,64,64,5,5,5) |           9257860          |           9522870       |  0.97X

Benchmarks are on Intel Skylake server (24 cores).

The patch also contains minor code refactor so that common helper class and function can be used by gemm_pack_rhs for both SpatialConvolution and CuboidConvolution.