@anuj-rawat 2019-06-25
> @anuj-rawat please sign CLA

Hi. I should already be on the CLA via Intel. In fact I have submitted patch to TensorFlow before #28205.
Is it possible that this CLA check failure is erroneous?