@anuj-rawat 2019-07-02
> Just noticed that some of the lines are waaaaay over 80 columns, could you please reformat all the changed files.

Sorry. I forgot to run `eigen_cuboid_convolution.h` through `clang-format`. I have now run all 3 `.h` files through `clag-format`.