@anuj-rawat 2019-07-02
> It fails presubmits with `./third_party/tensorflow/core/kernels/eigen_cuboid_convolution.h:561:17: error: unused variable 'packetSize' [-Werror,-Wunused-variable] const Index packetSize = internal::unpacket_traits<Packet>::size;`

I have removed the unused variable now. Sorry about that. The default build was not making unused variable warning into error.